requirements.txt
asgiref==3.4.1
blessings==1.7
bpython==0.21
certifi==2021.10.8
charset-normalizer==2.0.7
curtsies==0.3.10
cwcwidth==0.1.5
Django==3.2.8
django-crispy-forms==1.13.0
django-phonenumber-field==5.2.0
django-tenants==3.3.4
greenlet==1.1.2
idna==3.3
psycopg2==2.8.6
Pygments==2.10.0
pytz==2021.3
pyxdg==0.27
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
